add design

+++ lib/features/category/categories_bloc/categories_bloc.dart
@@ -19,8 +19,6 @@ class CategoriesBloc extends Bloc<CategoriesEvent, CategoriesState> {
     on<AddExpenseButtonPressedEvent>(_onAddExpensePressed);
     on<LoadCategoriesEvent>(_onLoadCategories);
     on<DeleteTransactionEvent>(_onDeleteTransaction);
-    on<EditTransactionEvent>(_onEditTransaction);
-    on<UpdateTransactionEvent>(_onUpdateTransaction);
   }
 
   Future<void> _onCategorySelected(
@@ -36,13 +34,15 @@ class CategoriesBloc extends Bloc<CategoriesEvent, CategoriesState> {
           .where('category', isEqualTo: event.category.name)
           .get();
 
-      final transactions = querySnapshot.docs
+      final transactions =
+      querySnapshot.docs
           .map((doc) => CategoryTransactionDtoFirestore.fromFirestore(doc))
           .toList();
 
       for (var t in transactions) {
-        debugPrint('Transaction id: ${t.id}, title: ${t.title}, category: ${t
-            .category}');
+        debugPrint(
+          'Transaction id: ${t.id}, title: ${t.title}, category: ${t.category}',
+        );
       }
 
       transactions.sort((a, b) => b.timeAndDate!.compareTo(a.timeAndDate!));
@@ -65,10 +65,8 @@ class CategoriesBloc extends Bloc<CategoriesEvent, CategoriesState> {
     emit(CategoriesInitialState());
   }
 
-  void _onAddExpensePressed(
-    AddExpenseButtonPressedEvent event,
-    Emitter<CategoriesState> emit,
-  ) {
+  void _onAddExpensePressed(AddExpenseButtonPressedEvent event,
+      Emitter<CategoriesState> emit,) {
     emit(CategoriesAddExpenseState());
   }
 
@@ -79,24 +77,22 @@ class CategoriesBloc extends Bloc<CategoriesEvent, CategoriesState> {
     try {
       final querySnapshot = await _firestore.collection('transactions').get();
 
-
-      final List<CategoryTransactionDto> transactions = querySnapshot.docs
+      final List<CategoryTransactionDto> transactions =
+      querySnapshot.docs
           .map((doc) {
         final data = doc.data();
 
-
-        if (data['title'] == null || data['category'] == null ||
+        if (data['title'] == null ||
+            data['category'] == null ||
             data['amount'] == null) {
           return null;
         }
-
 
         return CategoryTransactionDtoFirestore.fromFirestore(doc);
       })
           .where((dto) => dto != null)
           .cast<CategoryTransactionDto>()
           .toList();
-
 
       transactions.sort((a, b) => b.timeAndDate!.compareTo(a.timeAndDate!));
 
@@ -116,52 +112,26 @@ class CategoriesBloc extends Bloc<CategoriesEvent, CategoriesState> {
 
   Future<void> _onDeleteTransaction(DeleteTransactionEvent event,
       Emitter<CategoriesState> emit,) async {
+    if (state is! CategoriesLoadedState) return;
+
+    final currentState = state as CategoriesLoadedState;
+
+    emit(const CategoriesState.loading());
+
     try {
       await _firestore.collection('transactions').doc(event.id).delete();
 
-      add(
-        CategorySelectedEvent(
-          state is CategoriesLoadedState
-              ? (state as CategoriesLoadedState).selectedCategory
-              : CategoryEnum.more,
-          state is CategoriesLoadedState
-              ? (state as CategoriesLoadedState).selectedIndex
-              : -1,
+
+      final updatedTransactions = currentState.filteredTransactions
+          .where((t) => t.id != event.id)
+          .toList();
+
+      emit(
+        currentState.copyWith(
+          filteredTransactions: updatedTransactions,
         ),
       );
-    } catch (e, s) {
-      debugPrint('Error deleting transaction: $e');
-      debugPrintStack(stackTrace: s);
-      emit(CategoriesState.failure(e.toString()));
-    }
-  }
-
-  void _onEditTransaction(EditTransactionEvent event,
-      Emitter<CategoriesState> emit,) {
-    emit(CategoriesState.editExpense(event.transaction));
-  }
-
-  Future<void> _onUpdateTransaction(UpdateTransactionEvent event,
-      Emitter<CategoriesState> emit,) async {
-    try {
-      await _firestore
-          .collection('transactions')
-          .doc(event.id)
-          .update(event.updatedData);
-
-      add(
-        CategorySelectedEvent(
-          state is CategoriesLoadedState
-              ? (state as CategoriesLoadedState).selectedCategory
-              : CategoryEnum.more,
-          state is CategoriesLoadedState
-              ? (state as CategoriesLoadedState).selectedIndex
-              : -1,
-        ),
-      );
-    } catch (e, s) {
-      debugPrint('Error updating transaction: $e');
-      debugPrintStack(stackTrace: s);
+    } catch (e) {
       emit(CategoriesState.failure(e.toString()));
     }
   }
